Refactor controllers and enhance application structure
- Updated AlbumController and ArtistController to redirect to a new dashboard URL after CRUD operations.
- Modified AuthController to initialize context and added new authentication methods.
- Enhanced MusicController with related data fetching and added a Dashboard action.
- Updated project file to include new package references for DotNetEnv and Isopoh.Cryptography.Argon2.
- Refactored GlobalConfiguration to load settings from environment variables.
- Updated FloatlyContext with new entities and changed primary key types from long to int.
- Added new utility classes: EmailService, HashHelper, and StringExt.
- Created a new Dashboard interface in Dashboard.cshtml for management features.
- Added libman.json for managing client-side library dependencies.
- Introduced ApiController and versioned library endpoints for song and artist searches.
- Completely rewrote Login.cshtml and Index.cshtml for improved UI.
- Updated Program.cs to load environment variables and configure the database connection.

+++ Models/FormModel.cs
@@ -2,9 +2,9 @@
 {
     public class SongUploadModel
     {
-        public long? Id { get; set; }
+        public int? Id { get; set; }
         public string Title { get; set; } = null!;
-        public long? AlbumId { get; set; }
+        public int? AlbumId { get; set; }
 
         public IFormFile? MusicFile { get; set; }
         public IFormFile? LyricsFile { get; set; }
